refactor(opal): Inline presets, optional props, and edit UX improvements
- Delete presets.ts; inline HeadingLayout config into HeadingLayout.tsx
  and shared types into components.tsx (eliminates confusing unused entries)
- Make LabelLayout description optional (was required)
- Auto-select text when entering edit mode in both HeadingLayout and LabelLayout
- Update README with separate HeadingLayout/LabelLayout preset tables
