+++ web/lib/opal/src/components/layouts/Content/README.md
@@ -8,31 +8,38 @@ A two-axis layout component for displaying icon + title + description rows. Rout
 
 ### `sizePreset` — controls sizing (icon, padding, gap, font)
 
-| Preset | Icon | Container padding | Gap | Title font | Line-height |
+#### HeadingLayout presets
+
+| Preset | Icon | Icon padding | Gap | Title font | Line-height |
 |---|---|---|---|---|---|
 | `headline` | 2rem (32px) | `p-0.5` (2px) | 0.25rem (4px) | `font-heading-h2` | 2.25rem (36px) |
 | `section` | 1.25rem (20px) | `p-1` (4px) | 0rem | `font-heading-h3` | 1.75rem (28px) |
-| `main-content` | 1.125rem | `p-[0.1875rem]` | 0.125rem | `font-main-content-emphasis` | 1.5rem |
-| `main-ui` | 1rem | `p-0.5` | 0.25rem | `font-main-ui-action` | 1.25rem |
-| `secondary` | 0.75rem | `p-0.5` | 0.125rem | `font-secondary-action` | 1rem |
 
-> Icon container height (icon + 2 × padding) always equals the title line-height.
+#### LabelLayout presets
+
+| Preset | Icon | Icon padding | Icon color | Gap | Title font | Line-height |
+|---|---|---|---|---|---|---|
+| `main-content` | 1rem (16px) | `p-1` (4px) | `text-04` | 0.125rem (2px) | `font-main-content-emphasis` | 1.5rem (24px) |
+| `main-ui` | 1rem (16px) | `p-0.5` (2px) | `text-03` | 0.25rem (4px) | `font-main-ui-action` | 1.25rem (20px) |
+| `secondary` | 0.75rem (12px) | `p-0.5` (2px) | `text-04` | 0.125rem (2px) | `font-secondary-action` | 1rem (16px) |
+
+> Icon container height (icon + 2 x padding) always equals the title line-height.
 
 ### `variant` — controls structure / layout
 
 | variant | Description |
 |---|---|
 | `heading` | Icon on **top** (flex-col) — HeadingLayout |
-| `section` | Icon **inline** (flex-row) — HeadingLayout (or LabelLayout for smaller presets) |
+| `section` | Icon **inline** (flex-row) — HeadingLayout or LabelLayout |
 | `body` | Body text layout — BodyLayout (future) |
 
-### Valid Combinations → Internal Routing
+### Valid Combinations -> Internal Routing
 
 | sizePreset | variant | Routes to |
 |---|---|---|
 | `headline` / `section` | `heading` | **HeadingLayout** (icon on top) |
 | `headline` / `section` | `section` | **HeadingLayout** (icon inline) |
-| `main-content` / `main-ui` / `secondary` | `section` | LabelLayout (future) |
+| `main-content` / `main-ui` / `secondary` | `section` | **LabelLayout** |
 | `main-content` / `main-ui` / `secondary` | `body` | BodyLayout (future) |
 
 Invalid combinations (e.g. `sizePreset="headline" + variant="body"`) are excluded at the type level.
@@ -41,26 +48,23 @@ Invalid combinations (e.g. `sizePreset="headline" + variant="body"`) are exclude
 
 | Prop | Type | Default | Description |
 |---|---|---|---|
-| `sizePreset` | `SizePreset` | `"headline"` | Size preset (see table above) |
+| `sizePreset` | `SizePreset` | `"headline"` | Size preset (see tables above) |
 | `variant` | `ContentVariant` | `"heading"` | Layout variant (see table above) |
 | `icon` | `IconFunctionComponent` | — | Optional icon component |
 | `title` | `string` | **(required)** | Main title text |
-| `description` | `string` | — | Description below the title |
+| `description` | `string` | — | Optional description below the title |
 | `editable` | `boolean` | `false` | Enable inline editing of the title |
 | `onTitleChange` | `(newTitle: string) => void` | — | Called when user commits an edit |
 
-## Architecture
+## Internal Layouts
 
-```
-div.opal-content-heading                          <- outer flex (row or col)
-  +- div.opal-content-heading-icon-container      <- sized icon wrapper (centered)
-  |    +- Icon                                    (sized per sizePreset)
-  +- div.opal-content-heading-body                <- flex-1 column
-       +- div.opal-content-heading-title-row      <- flex row (title + edit button)
-       |    +- span.opal-content-heading-title    (or input when editing)
-       |    +- div.opal-content-heading-edit-button
-       +- div.opal-content-heading-description    <- secondary body text
-```
+### HeadingLayout
+
+For `headline` / `section` presets. Supports `variant="heading"` (icon on top) and `variant="section"` (icon inline). Description is always `font-secondary-body text-text-03`.
+
+### LabelLayout
+
+For `main-content` / `main-ui` / `secondary` presets. Always inline. Both `icon` and `description` are optional. Description is always `font-secondary-body text-text-03`.
 
 ## Usage Examples
 
@@ -68,7 +72,7 @@ div.opal-content-heading                          <- outer flex (row or col)
 import { Content } from "@opal/components";
 import SvgSearch from "@opal/icons/search";
 
-// Headline heading — large, icon on top
+// HeadingLayout — headline, icon on top
 <Content
   icon={SvgSearch}
   sizePreset="headline"
@@ -77,31 +81,27 @@ import SvgSearch from "@opal/icons/search";
   description="Configure your agent's behavior"
 />
 
-// Headline section — large, icon inline
-<Content
-  icon={SvgSearch}
-  sizePreset="headline"
-  variant="section"
-  title="Agent Settings"
-  description="Configure your agent's behavior"
-/>
-
-// Section heading — medium, icon on top
-<Content
-  icon={SvgSearch}
-  sizePreset="section"
-  variant="heading"
-  title="Data Sources"
-  description="Connected integrations"
-/>
-
-// Section section — medium, icon inline
+// HeadingLayout — section, icon inline
 <Content
   icon={SvgSearch}
   sizePreset="section"
   variant="section"
   title="Data Sources"
   description="Connected integrations"
+/>
+
+// LabelLayout — with icon and description
+<Content
+  icon={SvgSearch}
+  sizePreset="main-ui"
+  title="Instructions"
+  description="Agent system prompt"
+/>
+
+// LabelLayout — title only (no icon, no description)
+<Content
+  sizePreset="main-content"
+  title="Featured Agent"
 />
 
 // Editable title
